notification_service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Afsheen Shafiq <fsheekhi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.20"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pytest = "^8.2.2"
httpx = "^0.27.0"
aiokafka = "^0.11.0"
protobuf = "5.28.1"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
aiohttp = "^3.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
